refactor: 기능별/어플리케이션 단계별 패키지 분리
