+++ src/main/java/lotto/data/Lotto.java
@@ -1,4 +1,4 @@
-package lotto;
+package lotto.data;
 
 import java.util.List;
 import java.util.stream.Collectors;
