+++ src/main/java/lotto/purchaseresult/PurchaseResultProcessor.java
@@ -1,4 +1,7 @@
-package lotto;
+package lotto.purchaseresult;
+
+import lotto.data.Lotto;
+import lotto.data.PurchaseResult;
 
 import java.util.ArrayList;
 import java.util.Collections;
