+++ src/main/java/lotto/userinput/UserInputGetter.java
@@ -1,4 +1,4 @@
-package lotto;
+package lotto.userinput;
 
 import static camp.nextstep.edu.missionutils.Console.readLine;
